+++ lib/pages/iptv/iptv_player_page.dart
@@ -17,7 +17,6 @@ import '../../services/playback_coordinator.dart';
 import '../../services/player/player_settings.dart';
 import '../../services/window/window_service.dart';
 import '../../services/cast/cast_service.dart';
-import '../../services/discord/discord_rpc_service.dart';
 import '../../widgets/player/player_cast_sheet.dart';
 import '../../widgets/player/player_glass.dart';
 import '../../widgets/player/player_aspect_menu.dart';
@@ -204,7 +203,6 @@ class _IptvPlayerPageState extends State<IptvPlayerPage>
     }
     SystemChrome.setEnabledSystemUIMode(SystemUiMode.edgeToEdge);
     SystemChrome.setPreferredOrientations(DeviceOrientation.values);
-    DiscordRpcService.instance.clearToIdle();
     super.dispose();
   }
 
@@ -220,10 +218,6 @@ class _IptvPlayerPageState extends State<IptvPlayerPage>
 
   Future<void> _initPlayer() async {
     final myGeneration = ++_initGeneration;
-    DiscordRpcService.instance.setWatchingLiveTv(
-      channelName: widget.channel.name,
-      logoUrl: widget.channel.iconUrl,
-    );
     if (widget.hits.isEmpty) {
       setState(() {
         _isLoading = false;